feat/water-2644 -- controller moved to KPI reports api
