+++ src/lib/CRM.js
@@ -189,7 +189,7 @@ async function createColleague (request, h) {
  * returns the number of delegated access granted by month for the current year including the % change by month
  * with the totals for the year to date and overall total
  */
-const getKPIAccessRequests = async (request, h) => {
+const getEntityRolesKPIdata = async (request, h) => {
   const query = `
   SELECT date_part('month', created_at)::integer AS month,
   date_part('year', created_at)::integer as year,
@@ -203,28 +203,8 @@ const getKPIAccessRequests = async (request, h) => {
   if (!data) {
     return Boom.notFound('entity roles data not found', error);
   }
-  // sort the data in ascending order
-  const sorted = data.sort((a, b) => a.year + '' + a.month - b.year + '' + b.month);
-  // map the monthly data and calculate the percentage change by  month
-  const monthly = sorted.reduce((acc, row, index) => {
-    if (row.current_year) {
-      acc.push({
-        month: row.month,
-        total: row.total,
-        change: (row.total - sorted[(index - 1)].total) / sorted[(index - 1)].total * 100
-      });
-    }
-    return acc;
-  }, []);
 
-  // calculate the overall totals
-  const totals = data.reduce((acc, row) => {
-    acc.allTime = acc.allTime + row.total;
-    acc.ytd = row.current_year ? acc.ytd + row.total : acc.ytd;
-    return acc;
-  }, { allTime: 0, ytd: 0 });
-
-  return { data: { totals, monthly }, error: null };
+  return data;
 };
 
 module.exports = {
@@ -232,5 +212,5 @@ module.exports = {
   getColleagues,
   deleteColleague,
   createColleague,
-  getKPIAccessRequests
+  getEntityRolesKPIdata
 };
